Stop hiding untracked files in status
Since aryelgois/vcsh-mr contains a script to bootstrap
a new vcsh repository (called ,vcsh-init) and a '*' is
added by default on every gitignore file, it is safe
and useful to list untracked files.

When I want to add a new file, first I must 'unignore'
all directories from $HOME to the file, then the file
itself.